Meson build system support (#159)
This adds `meson` as an alternative build system.

There's only a single configuration option, `build_tests`, that controls
whether the tests are built, enabled by default and used only if the
project is not built as a sub-project to mirror the CMake behavior.

When the tests are enabled, pulls `googletest` either from the system or
from source release with WrapDB patches (see
https://mesonbuild.com/Wrapdb-projects.html) - if needed, that can be
changed by modifying the `suprojects/gtest.wrap` file.

---------

Co-authored-by: Ilya Andreev <bugdealer@icloud.com>